Added tests to TestMountainClimbing.java.
Fixed typo in MountainClimbing.java.

+++ main/DynamicProgramming/MountainClimbing.java
@@ -3,24 +3,24 @@ import java.util.*;
 
 public class MountainClimbing
 {
-	// takes a 2d array of ints, will return the largest possible path travleing up the mountain
+	// takes a 2d array of longs, will return the largest possible path traveling up the mountain
 	// when you can move up, up-right, and up-left, the mountain is expected to be a rectangle
 	// the last array is assumed to be the top of the mountain
 	// all value of the are assumed to be positive, however the algorithm can be easily adapted to accomodate negatives
-	public int climbing(int[][] mountain)
+	public long climbing(long[][] mountain)
 	{
-		int height = mountain.length;
-		int width = mountain[0].length;
+		long height = mountain.length;
+		long width = mountain[0].length;
 
-		int[] cur_best = mountain[height-1];
+		long[] cur_best = mountain[height-1];
 
-		int down = 0;
-		int down_left = 0;
-		int down_right = 0;
-		for (int h = 1; h < height; h++)
+		long down = 0;
+		long down_left = 0;
+		long down_right = 0;
+		for (long h = 1; h < height; h++)
 		{
-			int[] new_best = mountain[h];
-			for (int w = 0; w < width; w++)
+			long[] new_best = mountain[h];
+			for (long w = 0; w < width; w++)
 			{
 				down = cur_best[w];
 				down_right = (w < width-2) ? cur_best[w+1] : 0;
@@ -42,8 +42,8 @@ public class MountainClimbing
 			cur_best = new_best;
 		}
 
-		int max = 0;
-		for (int i = 0; i < width; i++)
+		long max = 0;
+		for (long i = 0; i < width; i++)
 		{
 			if (cur_best[i] > max)
 			{
